requirements.txt
Flask==2.1.3
pandas==1.3.5
pycryptodome==3.15.0
pyppeteer==1.0.2
regex==2022.1.18
requests==2.27.1
scipy==1.8.0
selenium==4.3.0
webdriver_manager==3.5.4
